refact : 사용자가 작성한 시음기록 조회 로직 이동
시음기록 도메인에서 관리하도록 수정
profiles -> records/tasted_records
