+++ repo/profiles/urls.py
@@ -17,6 +17,5 @@ urlpatterns = [
     path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path("<int:id>/posts/", views.UserPostListAPIView.as_view(), name="user_posts"),
-    path("<int:id>/tasted-records/", views.UserTastedRecordListView.as_view(), name="user_tasted_records"),
     path("<int:id>/notes/", views.UserNoteAPIView.as_view(), name="user_notes"),
 ]
